+++ app/src/main/java/es/chiteroman/playintegrityfix/EntryPoint.java
@@ -1,0 +1,85 @@
+package es.chiteroman.playintegrityfix;
+
+import android.os.Build;
+import android.util.Log;
+
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.lang.reflect.Field;
+import java.security.KeyStore;
+import java.security.KeyStoreException;
+import java.security.KeyStoreSpi;
+import java.security.Provider;
+import java.security.Security;
+import java.util.Properties;
+
+public final class EntryPoint {
+    private static final Properties props = new Properties();
+    private static final File file = new File("/data/data/com.google.android.gms/cache/pif.prop");
+
+    public static void init() {
+
+        try (FileInputStream inputStream = new FileInputStream(file)) {
+            props.load(inputStream);
+            LOG("Loaded " + props.size() + " properties!");
+        } catch (IOException e) {
+            LOG("Couldn't load properties file: " + e);
+            return;
+        }
+
+        spoofDevice();
+        spoofProvider();
+    }
+
+    private static void spoofProvider() {
+        final String KEYSTORE = "AndroidKeyStore";
+
+        try {
+            Provider provider = Security.getProvider(KEYSTORE);
+            KeyStore keyStore = KeyStore.getInstance(KEYSTORE);
+
+            Field f = keyStore.getClass().getDeclaredField("keyStoreSpi");
+            f.setAccessible(true);
+            CustomKeyStoreSpi.keyStoreSpi = (KeyStoreSpi) f.get(keyStore);
+            f.setAccessible(false);
+
+            CustomProvider customProvider = new CustomProvider(provider);
+            Security.removeProvider(KEYSTORE);
+            Security.insertProviderAt(customProvider, 1);
+
+            LOG("Spoof KeyStoreSpi and Provider done!");
+
+        } catch (KeyStoreException e) {
+            LOG("Couldn't find KeyStore: " + e);
+        } catch (NoSuchFieldException e) {
+            LOG("Couldn't find field: " + e);
+        } catch (IllegalAccessException e) {
+            LOG("Couldn't change access of field: " + e);
+        }
+    }
+
+    public static void spoofDevice() {
+        props.forEach((field, value) -> setProp((String) field, (String) value));
+    }
+
+    private static void setProp(String name, String value) {
+        try {
+            Field field = Build.class.getDeclaredField(name);
+            field.setAccessible(true);
+            String oldValue = (String) field.get(null);
+            field.set(null, value);
+            field.setAccessible(false);
+            if (value.equals(oldValue)) return;
+            LOG(String.format("[%s]: %s -> %s", name, oldValue, value));
+        } catch (NoSuchFieldException e) {
+            LOG(String.format("Couldn't find '%s' field name.", name));
+        } catch (IllegalAccessException e) {
+            LOG(String.format("Couldn't modify '%s' field value.", name));
+        }
+    }
+
+    public static void LOG(String msg) {
+        Log.d("PIF/Java", msg);
+    }
+}