+++ changelog.md
@@ -1,5 +1,11 @@
-# v13
+# v13.1
 
-Device verdict is green again x2.
+For normal users:
+- Custom props! You can edit them in file 'pif.prop'
 
-If you are using xiaomi.eu expect to fail all tests.
+For devs:
+- Source code published again.
+- Drop support for emulators (x86 and x86_64).
+- Use ShadowHook instead Dobby.
+- Updated libcxx.
+- 'classes.dex' is back to module folder instead hardcoding it in libs.